fix(v-slider): added missing click events for prepend and append icons
There were props for adding click events for the append and prepend icons in the v slider but they
weren't used in the component which led to not being able to have events for the icons. I also added
unit tests that check if the icons are clicked they have emitted the event.

fix #16071